added looksrare sales scraping

+++ main.ts
@@ -15,9 +15,10 @@ import fs from 'fs';
 import { Database } from 'sqlite3';
 import dotenv from 'dotenv';
 
+
 /// Use this if you wanna force recreation the initial database
 const REGENERATE_FROM_SCRATCH = false;
-const CHUNK_SIZE = 500; // lower this if geth node is hanging
+const CHUNK_SIZE = 200; // lower this if geth node is hanging
 const RARIBLE_TOPIC0 = '0xcae9d16f553e92058883de29cb3135dbc0c1e31fd7eace79fef1d80577fe482e';
 const NFTX_TOPIC0 = '0xf7735c8cb2a65788ca663fc8415b7c6a66cd6847d58346d8334e8d52a599d3df';
 const NFTX_ALTERNATE_TOPIC0 = '0x1cdb5ee3c47e1a706ac452b89698e5e3f2ff4f835ca72dde8936d0f4fcf37d81';
@@ -26,6 +27,7 @@ const NFTX_SELL_TOPIC0 = '0x1cdb5ee3c47e1a706ac452b89698e5e3f2ff4f835ca72dde8936
 const CARGO_TOPIC0 = '0x5535fa724c02f50c6fb4300412f937dbcdf655b0ebd4ecaca9a0d377d0c0d9cc';
 const PHUNK_MARKETPLACE_TOPIC0 = '0x975c7be5322a86cddffed1e3e0e55471a764ac2764d25176ceb8e17feef9392c';
 const OPENSEA_SALE_TOPIC0 = '0xc4109843e0b7d514e4c093114b863f8e7d8d9a458c372cd51bfe526b588006c9';
+const LOOKSRARE_SALE_TOPIC0 = '0x95fb6205e23ff6bda16a2d1dba56b9ad7c783f67c96fa149785052f47696f2be';
 
 if (fs.existsSync('.env.local')) {
   dotenv.config({ path: '.env.local' });
@@ -35,7 +37,7 @@ if (fs.existsSync('.env.local')) {
 }
 
 const readFile = promisify(fs.readFile);
-console.log(`opening database at ${process.env.WORK_DIRECTORY + process.env.DATABASE_FILE}`)
+console.log(`opening database at ${process.env.WORK_DIRECTORY + process.env.DATABASE_FILE}`);
 let db = new Database(process.env.WORK_DIRECTORY + process.env.DATABASE_FILE);
 
 async function work() {
@@ -94,7 +96,8 @@ async function work() {
             || l.topics[0] === NFTX_ALTERNATE_TOPIC0
             || l.topics[0] === CARGO_TOPIC0
             || l.topics[0] === PHUNK_MARKETPLACE_TOPIC0
-            || l.topics[0] === OPENSEA_SALE_TOPIC0) {
+            || l.topics[0] === OPENSEA_SALE_TOPIC0
+            || l.topics[0] === LOOKSRARE_SALE_TOPIC0) {
             saleFound = true;
           }
           if (l.topics[0] === OPENSEA_SALE_TOPIC0) {
@@ -125,6 +128,27 @@ async function work() {
               console.log('already exist! we have to debug that!');
             }
             console.log(`\n${txDate.toLocaleString()} - indexed an opensea sale for token #${tokenId} to 0x${targetOwner} for ${web3.utils.fromWei(amount.toString(), 'ether')}eth in tx ${tr.transactionHash}.`);
+          } else if (l.topics[0] === LOOKSRARE_SALE_TOPIC0) {
+            const data = l.data.substring(2);
+            const dataSlices = data.match(/.{1,64}/g);
+            const amount = parseInt(dataSlices[6], 16);
+            const tokenId = ev.returnValues.tokenId;
+            const targetOwner = ev.returnValues.to.toLowerCase();
+            const sourceOwner = ev.returnValues.from.toLowerCase();
+            const rowExists = await new Promise((resolve) => {
+              db.get('SELECT * FROM events WHERE tx = ? AND log_index = ?', [ev.transactionHash, ev.logIndex], (err, row) => {
+                if (err) {
+                  resolve(false);
+                }
+                resolve(row !== undefined);
+              });
+            });
+            if (!rowExists) {
+              const stmt = db.prepare('INSERT INTO events VALUES (?,?,?,?,?,?,?,?,?)');
+              stmt.run('sale', sourceOwner, targetOwner, tokenId, parseFloat(new BN(amount.toString()).toString()), txDate.toISOString(), ev.transactionHash, ev.logIndex, 'looksrare');
+              stmt.finalize();
+            }
+            console.log(`\n${txDate.toLocaleString()} - indexed a looksrare sale for token #${tokenId} to ${targetOwner} for ${web3.utils.fromWei(amount.toString(), 'ether')}eth in tx ${tr.transactionHash}.`);            
           } else if (l.topics[0] === PHUNK_MARKETPLACE_TOPIC0) {
             const data = l.data.substring(2);
             const dataSlices = data.match(/.{1,64}/g);
@@ -353,7 +377,7 @@ async function createDatabaseIfNeeded() {
   });
   if (REGENERATE_FROM_SCRATCH || !tableExists) {
     console.log('Recreating database...');
-    fs.unlinkSync(process.env.DATABASE_FILE);
+    if (fs.existsSync(process.env.DATABASE_FILE)) fs.unlinkSync(process.env.DATABASE_FILE);
     db = new Database(process.env.WORK_DIRECTORY + process.env.DATABASE_FILE);
     db.serialize(() => {
       console.log('create table');
